create logging tool
