+++ tools/logging/main.go
@@ -1,0 +1,66 @@
+package logging
+
+import (
+	"fmt"
+	"os"
+	"strings"
+	"time"
+
+	"github.com/rs/zerolog"
+	"github.com/rs/zerolog/log"
+)
+
+// colorize returns the string s wrapped in ANSI code c, unless disabled is true.
+func colorize(s interface{}, c int) string {
+	return fmt.Sprintf("\x1b[%dm%v\x1b[0m", c, s)
+}
+
+func formatLevel(i interface{}) string {
+	var l string
+	if ll, ok := i.(string); ok {
+		switch ll {
+		case "debug":
+			l = colorize(ll, colorBlue)
+		case "info":
+			l = colorize(ll, colorGreen)
+		case "warn":
+			l = colorize(ll, colorYellow)
+		case "error":
+			l = colorize(colorize(ll, colorRed), colorBold)
+		case "fatal":
+			l = colorize(colorize(ll, colorRed), colorBold)
+		case "panic":
+			l = colorize(colorize(ll, colorRed), colorBold)
+		default:
+			l = colorize("???", colorBold)
+		}
+	} else {
+		if i == nil {
+			l = colorize("???", colorBold)
+		} else {
+			l = strings.ToUpper(fmt.Sprintf("%s", i))[0:3]
+		}
+	}
+	return l
+}
+
+func SetupLogging(level string) {
+	if level == "debug" {
+		zerolog.SetGlobalLevel(zerolog.DebugLevel)
+	} else if level == "info" {
+		zerolog.SetGlobalLevel(zerolog.InfoLevel)
+	} else if level == "warn" {
+		zerolog.SetGlobalLevel(zerolog.WarnLevel)
+	} else if level == "error" {
+		zerolog.SetGlobalLevel(zerolog.ErrorLevel)
+	} else if level == "fatal" {
+		zerolog.SetGlobalLevel(zerolog.FatalLevel)
+	} else if level == "panic" {
+		zerolog.SetGlobalLevel(zerolog.PanicLevel)
+	}
+	log.Logger = log.Output(zerolog.ConsoleWriter{
+		TimeFormat:  time.RFC3339,
+		Out:         os.Stdout,
+		FormatLevel: formatLevel,
+	})
+}
